torrent-ui/src/main/java/com/gorolykmaxim/torrentui/common/keyvalue/KeyValue.java
package com.gorolykmaxim.torrentui.common.keyvalue;

import javax.persistence.Entity;
import javax.persistence.Id;
import java.util.Objects;

@Entity
public class KeyValue {
    @Id
    private String key;
    private String value;

    protected KeyValue() {}

    public KeyValue(String key, String value) {
        this.key = key;
        this.value = value;
    }

    public String getKey() {
        return this.key;
    }

    public String getValue() {
        return this.value;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        KeyValue keyValue = (KeyValue) o;
        return Objects.equals(key, keyValue.key) &&
                Objects.equals(value, keyValue.value);
    }

    @Override
    public int hashCode() {
        return Objects.hash(key, value);
    }
}
